Home styling (sort of) done!
Finally fixed the weird margin issue and ended up sticking with << and >> month-time navigation.  I played with adding a swipe gesture for the month calendar to navigate in time, but it didn't feel very intuitive, and realistically you don't need to swipe quickly between many months at once.  Instead, it'd probably be better to build a "year in review" view or some other summary view.

- Increased margin size for MonthVC so we don't end up with weird scrunching
- Changed Home nav options / going to keep them as is for v1