Home (Month/Day) view layout tweaks to success!
Finally tweaked the math a bit to ensure that the cells fit 7 per row on iPhone 4s/5/6/6p.  This took a bit of a hack in rounding the minInteritemSpacing down to the nearest point, in that I suspect when it was calculated exactly, overall the width would be too wide and was squeezing cells down into additional rows.  Basically, on iPhone6, it'd end up rendering 6 cells per row instead 7.

- Tweaked spacing math for MonthViewController
- Minor cleanup of code for HomeViewController
- 1 hack to set the month's frame size explicitly -- because of where the estimated sizing was being down in the VC lifecycle, the width from self.frame was off
- another hack to increase the estimated height by ~1/2 a cell width to give some buffer room and ensure the bottommost row of cells wasn't being cut off